+++ H/ugens3.h
@@ -26,7 +26,7 @@
 typedef struct {
         OPDS    h;
         MYFLT   *rslt, *xamp, *kcps, *xcar, *xmod, *kndx, *ifn, *iphs;
-        long    mphs, cphs;
+        int32    mphs, cphs;
         short ampcod, carcod, modcod;
         FUNC    *ftp;
 } FOSC;
@@ -35,10 +35,10 @@ typedef struct {
         OPDS    h;
         MYFLT   *ar1,*ar2,*xamp,*kcps,*ifn,*ibas,*imod1,*ibeg1,*iend1,*imod2,*ibeg2,*iend2;
         MYFLT   cpscvt;
-        long    lphs;
+        int32    lphs;
         short   mod1, mod2;
-        long    beg1, beg2;
-        long    end1, end2;
+        int32    beg1, beg2;
+        int32    end1, end2;
         short   seg1, curmod, looping, stereo;
         FUNC    *ftp;
 } LOSC;
@@ -53,7 +53,7 @@ typedef struct ptlptr {
         DUPLE   *ap;
         DUPLE   *fp;
         short   amp,frq;
-        long    phs;
+        int32    phs;
 } PTLPTR;
 
 #define MAXPTLS 50    /* must agree with hetro.c */
@@ -62,6 +62,6 @@ typedef struct {
         OPDS    h;
         MYFLT   *rslt, *kamod, *kfmod, *ksmod, *ifilcod, *dum;
         MEMFIL  *mfp;
-        long    mksecs;
+        int32   mksecs;
         AUXCH   aux;            /* PTLPTR  ptlptrs[MAXPTLS + 1]; make dynamic */
 } ADSYN;
