+++ H/ugens4.h
@@ -27,7 +27,7 @@ typedef struct {
         OPDS    h;
         MYFLT   *ar, *xamp, *xcps, *knh, *ifn, *iphs;
         short   ampcod, cpscod;
-        long    lphs;
+        int32    lphs;
         FUNC    *ftp;
         int     reported;
 } BUZZ;
@@ -37,7 +37,7 @@ typedef struct {
         MYFLT   *ar, *xamp, *xcps, *kn, *kk, *kr, *ifn, *iphs;
         short   ampcod, cpscod, prvn;
         MYFLT   prvr, twor, rsqp1, rtn, rtnp1, rsumr;
-        long    lphs;
+        int32    lphs;
         FUNC    *ftp;
         int     reported;
         MYFLT   last;
@@ -48,7 +48,7 @@ typedef struct {
         MYFLT   *ar, *kamp, *kcps, *icps, *ifn, *imeth, *ipar1, *ipar2;
         MYFLT   sicps, param1, param2;
         short   thresh1, thresh2, method;
-        long    phs256, npts, maxpts;
+        int32    phs256, npts, maxpts;
         AUXCH   auxch;
 } PLUCK;
 
@@ -65,7 +65,7 @@ typedef struct {
         MYFLT   *ar, *xamp, *xcps, *iseed, *sel, *base;
         short   ampcod, cpscod, new;
         int     rand;
-        long    phs;
+        int32    phs;
         MYFLT   num1;
 } RANDH;
 
@@ -74,6 +74,6 @@ typedef struct {
         MYFLT   *ar, *xamp, *xcps, *iseed, *sel, *base;
         short   ampcod, cpscod, new;
         int     rand;
-        long    phs;
+        int32    phs;
         MYFLT   num1, num2, dfdmax;
 } RANDI;
